+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Java 9+ support
 - Chinese translation updated
 - Enable bulk replace for single typed tags
+- Option to use logging configuration file (ffdec home / logging.properties)
 
 ### Removed
 - Due to Java9+ limited support of reflection, following features were removed: